Migrate UsdMedia to views; add sdf::TimeCode
Convert the media schema to typed Prim views mirroring geom/lux:
SpatialAudio (a geom::Xformable) and the AssetPreviewsAPI single-apply
view, dropping the read_*/define_*/Read* surface. media now enables the
geom feature, since SpatialAudio is UsdGeomXformable-derived.

Add sdf::TimeCode (= SdfTimeCode), a value newtype over f64 with
From/TryFrom<Value>, so timecode attributes read/write typed via
get::<sdf::TimeCode>() / set(sdf::TimeCode(..)).

Settle the schema-module convention: per family, mod.rs (glue + macro +
helpers + token enums), tokens.rs, schema.rs (grouped view structs), and
optional traits.rs. Fold geom's internal.rs into mod.rs, rename lux
lights.rs to schema.rs, merge shade preview.rs into types.rs, and drop
decorative box-drawing divider comments (with a CLAUDE.md rule).
