+++ ROADMAP.md
@@ -188,7 +188,7 @@ Features from the C++ reference implementation not covered by the core specifica
 | [UsdVol](https://openusd.org/release/api/usd_vol_page_front.html) (volumes) | :white_check_mark: | `main` | Reader + authoring behind `vol` feature: `Volume` (`field:<name>` relationships), `OpenVDBAsset` / `Field3DAsset` (shared FieldAsset attrs)<br>Remaining — the `ParticleField*` Gaussian-splat schemas |
 | [UsdPhysics](https://openusd.org/release/api/usd_physics_page_front.html) schema reader | :white_check_mark: | `0.4.0` | Reader behind `physics` feature: all 8 prim types, the 7 single-apply APIs, and the LimitAPI / DriveAPI multi-apply schemas<br>Remaining — authoring helpers; collection evaluation for `CollisionGroup` (tracked under Spec 15) |
 | [UsdRender](https://openusd.org/release/api/usd_render_page_front.html) | :white_check_mark: | `main` | Reader + authoring behind `render` feature: RenderSettingsBase / RenderSettings / RenderProduct / RenderVar / RenderPass, and the computed render spec (`compute_render_spec`) with product-overrides-settings inheritance, the aspect-ratio conform policy, render-var de-duplication, and per-level `namespacedSettings`<br>Remaining — RenderPass collection memberships (needs CollectionAPI evaluation); node-graph-driven namespaced settings (needs UsdShade value resolution); authoring `renderSettingsPrimPath` stage metadata; UsdRenderDenoisePass (dev schema) |
-| [UsdMedia](https://openusd.org/release/api/usd_media_page_front.html) | :white_check_mark: | `main` | Reader + authoring behind `media` feature: `SpatialAudio` (filePath / auralMode / playbackMode / startTime / endTime / mediaOffset / gain) and `AssetPreviewsAPI` (default thumbnail via `assetInfo.previews`) |
+| [UsdMedia](https://openusd.org/release/api/usd_media_page_front.html) | :white_check_mark: | `main` | Trait-views behind `media` feature (built on the `geom` chain): `SpatialAudio` (an `Xformable`: filePath / auralMode / playbackMode / startTime / endTime / mediaOffset / gain) and the applied `AssetPreviewsAPI` (default thumbnail via `assetInfo.previews`) |
 | [UsdUI](https://openusd.org/release/api/usd_u_i_page_front.html) | :white_check_mark: | `main` | Reader + authoring behind `ui` feature: `SceneGraphPrimAPI` (displayName / displayGroup), `NodeGraphNodeAPI` (node-editor layout), `Backdrop` |
 | [UsdProc](https://openusd.org/release/api/usd_proc_page_front.html) | :white_check_mark: | `main` | Reader + authoring behind `proc` feature: `GenerativeProcedural` (`proceduralSystem`) |
 | [Flatten / export](https://openusd.org/release/api/flatten_utils_8h.html) | :construction: | | |
