+++ src/schemas/geom/camera.rs
@@ -5,8 +5,8 @@ use anyhow::Result;
 use crate::sdf;
 use crate::usd::{Attribute, Prim, SchemaBase, SchemaKind, Stage};
 
-use super::internal::{create, create_uniform_token, get_typed};
 use super::tokens as tok;
+use super::{create, create_uniform_token, get_typed};
 use super::{impl_geom_schema, Imageable, Xformable};
 
 /// A camera prim (C++ `UsdGeomCamera`) — an [`Xformable`] exposing the lens /
