+++ src/schemas/geom/grouping.rs
@@ -5,7 +5,7 @@ use anyhow::Result;
 use crate::sdf;
 use crate::usd::{Prim, SchemaBase, SchemaKind, Stage};
 
-use super::internal::get_typed;
+use super::get_typed;
 use super::tokens as tok;
 use super::{impl_geom_schema, Imageable, Xformable};
