+++ src/schemas/geom/shapes.rs
@@ -13,8 +13,8 @@ use anyhow::Result;
 use crate::sdf;
 use crate::usd::{Attribute, Prim, SchemaBase, SchemaKind, Stage};
 
-use super::internal::{create, create_uniform_token, get_typed};
 use super::tokens as tok;
+use super::{create, create_uniform_token, get_typed};
 use super::{impl_geom_schema, Boundable, Gprim, Imageable, Xformable};
 
 /// A cube centered at the prim origin (C++ `UsdGeomCube`). Its one intrinsic
